requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2
django-anymail==9.1
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-environ==0.9.0
django-import-export==3.2.0
django-jazzmin==2.6.0
django-js-asset==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
environs==14.1.1
et_xmlfile==2.0.0
humanize==4.6.0
idna==3.10
inflection==0.5.1
marshmallow==3.26.1
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.2
pillow==11.2.1
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
setuptools==78.1.0
shortuuid==1.0.11
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
xlrd==2.0.1
xlwt==1.3.0
